fix ias in unit tests
